requirements.txt
fandom-py
requests
beautifulsoup4
customtkinter
